+++ app/src/main/java/com/techcourse/controller/ForwardController.java
@@ -1,12 +1,12 @@
-package webmvc.org.springframework.web.servlet.mvc.asis;
+package com.techcourse.controller;
 
 import context.org.springframework.stereotype.Controller;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import web.org.springframework.web.bind.annotation.RequestMapping;
 import web.org.springframework.web.bind.annotation.RequestMethod;
-import webmvc.org.springframework.web.servlet.ModelAndView;
-import webmvc.org.springframework.web.servlet.view.RedirectView;
+import webmvc.org.springframework.web.servlet.mvc.ModelAndView;
+import webmvc.org.springframework.web.servlet.mvc.view.RedirectView;
 
 @Controller(path = "/")
 public class ForwardController{
