+++ app/src/main/java/com/techcourse/controller/LoginController.java
@@ -9,8 +9,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import web.org.springframework.web.bind.annotation.RequestMapping;
 import web.org.springframework.web.bind.annotation.RequestMethod;
-import webmvc.org.springframework.web.servlet.ModelAndView;
-import webmvc.org.springframework.web.servlet.view.RedirectView;
+import webmvc.org.springframework.web.servlet.mvc.ModelAndView;
+import webmvc.org.springframework.web.servlet.mvc.view.RedirectView;
 
 @Controller(path = "/login")
 public class LoginController {
