+++ app/src/main/java/com/techcourse/controller/RegisterController.java
@@ -7,8 +7,8 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import web.org.springframework.web.bind.annotation.RequestMapping;
 import web.org.springframework.web.bind.annotation.RequestMethod;
-import webmvc.org.springframework.web.servlet.ModelAndView;
-import webmvc.org.springframework.web.servlet.view.JspView;
+import webmvc.org.springframework.web.servlet.mvc.ModelAndView;
+import webmvc.org.springframework.web.servlet.mvc.view.JspView;
 
 @Controller(path = "/register")
 public class RegisterController {
