+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/DispatcherServlet.java
@@ -9,10 +9,10 @@ import java.util.List;
 import java.util.Objects;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import webmvc.org.springframework.web.servlet.ModelAndView;
-import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerAdapter;
-import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerMapping;
-import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerNotFoundException;
+import webmvc.org.springframework.web.servlet.mvc.context.WebApplicationContext;
+import webmvc.org.springframework.web.servlet.mvc.handleradapter.HandlerAdapter;
+import webmvc.org.springframework.web.servlet.mvc.handlermapping.HandlerMapping;
+import webmvc.org.springframework.web.servlet.mvc.handlermapping.HandlerNotFoundException;
 
 public class DispatcherServlet extends HttpServlet {
