+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/handleradapter/HandlerAdapter.java
@@ -1,8 +1,8 @@
-package webmvc.org.springframework.web.servlet.mvc.tobe;
+package webmvc.org.springframework.web.servlet.mvc.handleradapter;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import webmvc.org.springframework.web.servlet.ModelAndView;
+import webmvc.org.springframework.web.servlet.mvc.ModelAndView;
 
 public interface HandlerAdapter {
