+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/handlermapping/AnnotationHandlerMapping.java
@@ -1,4 +1,4 @@
-package webmvc.org.springframework.web.servlet.mvc.tobe;
+package webmvc.org.springframework.web.servlet.mvc.handlermapping;
 
 import context.org.springframework.stereotype.Controller;
 import jakarta.servlet.http.HttpServletRequest;
@@ -14,7 +14,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import web.org.springframework.web.bind.annotation.RequestMapping;
 import web.org.springframework.web.bind.annotation.RequestMethod;
-import webmvc.org.springframework.web.servlet.mvc.WebApplicationContext;
+import webmvc.org.springframework.web.servlet.mvc.context.WebApplicationContext;
 
 public class AnnotationHandlerMapping implements HandlerMapping {
