+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/handlermapping/HandlerNotFoundException.java
@@ -1,4 +1,4 @@
-package webmvc.org.springframework.web.servlet.mvc.tobe;
+package webmvc.org.springframework.web.servlet.mvc.handlermapping;
 
 public class HandlerNotFoundException extends RuntimeException {
