--- mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/tobe/SimpleControllerHandlerAdapter.java
@@ -1,26 +1,0 @@
-package webmvc.org.springframework.web.servlet.mvc.tobe;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import webmvc.org.springframework.web.servlet.ModelAndView;
-import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
-import webmvc.org.springframework.web.servlet.view.JspView;
-
-public class SimpleControllerHandlerAdapter implements HandlerAdapter {
-
-    @Override
-    public boolean supports(final Object handler) {
-        return handler instanceof Controller;
-    }
-
-    @Override
-    public ModelAndView handle(
-            final HttpServletRequest request,
-            final HttpServletResponse response,
-            final Object handler
-    ) throws Exception {
-        final var viewName = ((Controller) handler).execute(request, response);
-        final var view = new JspView(viewName);
-        return new ModelAndView(view);
-    }
-}
