+++ mvc/src/test/java/webmvc/org/springframework/web/servlet/mvc/tobe/AnnotationHandlerMappingTest.java
@@ -8,7 +8,9 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import webmvc.org.springframework.web.servlet.mvc.WebApplicationContext;
+import webmvc.org.springframework.web.servlet.mvc.handlermapping.AnnotationHandlerMapping;
+import webmvc.org.springframework.web.servlet.mvc.handlermapping.HandlerExecution;
+import webmvc.org.springframework.web.servlet.mvc.context.WebApplicationContext;
 
 class AnnotationHandlerMappingTest {
